third_party/blink/web_tests/platform/generic/fast/forms/date-multiple-fields/date-multiple-fields-onblur-setvalue-onfocusremoved-expected.txt

Setting value onBlur, should not keep element in focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.getPropertyValue("background-color") is "rgb(0, 0, 255)"
PASS style.getPropertyValue("background-color") is "rgb(255, 255, 255)"
PASS successfullyParsed is true

TEST COMPLETE
